Refactor tests layout
