+++ test/clear.js
@@ -1,7 +1,7 @@
 'use strict'
 
 var createContext = require('./util/create-context')
-var createREGL = require('../../regl')
+var createREGL = require('../regl')
 var tape = require('tape')
 
 tape('clear color', function (t) {
