+++ test/framebuffer-codegen.js
@@ -1,5 +1,5 @@
 var createContext = require('./util/create-context')
-var createREGL = require('../../regl')
+var createREGL = require('../regl')
 var extend = require('../lib/util/extend')
 var tape = require('tape')
