+++ test/viewport.js
@@ -1,6 +1,6 @@
 var extend = require('../lib/util/extend')
 var createContext = require('./util/create-context')
-var createREGL = require('../../regl')
+var createREGL = require('../regl')
 var tape = require('tape')
 
 tape('viewport', function (t) {
